agent-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
fastapi-cors
python-multipart
ffmpeg-python
requests
websocket-client